feat(lazy): setup lazy and test auto-merge
Lazy merges all Lua scripts automatically in the plugins folder (only at
the root level of the plugins folder, it doesn't work in subfolders of
'lua/plugins').

The structure proposed in this commit is a test an probably temporary.
I'll make some changes in the future to improve this configuration.

Additionally, I've added the rose-pine theme and modularized the
colorscheme configuration as initial setup. I might consider leveraging
this structure to create a colorscheme update script in the future.

Signed-off-by: Carlos Sandoval <csandovalc@unal.edu.co>